+++ src/FaluSdk/PaymentRefunds/PaymentReversalMpesaDetails.cs
@@ -1,4 +1,4 @@
-﻿namespace Falu.PaymentReversals
+﻿namespace Falu.PaymentRefunds
 {
     /// <summary>
     /// Represents the details for an MPESA payment reversal.
